205 - Pager/Fetcher all the things! (#231)
Convert remaining models to user `Pager` or `Fetchers` for sub-items (eg views on workbook, workbooks on user, connections on data source). This works by wrapping an internal call to Server in a closure, then binding that to the item's property. Pretty simple. We first did this pattern for users in groups. Some lists aren't paged (eg connections) so I just made a simple fetch function and bind that. No need for a pager there.

This means that each call to `ModelItem.subItem` will fetch a new List/Pager from Server. This is a better pattern for staying up to date, and removes the need to synchronize things locally. If the caller needs a static list they can just assign it to a variable.

I left tags as they are -- that's a much bigger change that we need to think through.