Update Pager to handle un-paged lists (#322)
Not all endpoints support pagination on the server-side. We don't really give a way to predict what will or won't work with `Pager`, so I've updated it to check for a `total_available` amount. If that is `None` the endpoint doesn't support paging, and we can just drain the existing list and return. If the endpoint gets updated to support paging, it'll just magically start working.