Fixes #355
Included documentation on how to login to Tableau Online as well as Tableau Server. 

Also included a commented-out line to demonstrate how to specify a certain REST API version, which helps users maintain their code. Promoted using a "with" block to sign in since it is a best practice for managing files and requires fewer commands.