parse_datetime() should gracefully handle invalid dates (#529)
* parse_datetime() should gracefully handle invalid dates

We've seen instances where enumerating workbooks on our server results in an exception in parse_datetime(). It appears that some workbooks contain invalid dates e.g. `2018-06-31T13:12:00Z` for their created_at or updated_at times which causes the method to throw an exception, breaking enumeration.

This change simply catches the parsing failure and returns `None` in that scenario.

* Handle date is `None` in format_datetime()

* Add test XML file with invalid date

* Add test for invalid date

* Add missing constant

* Remove whitespace