Expose the `fileuploads` API endpoint
We had at least two independent re-implementations [1, 2] of file uploads
within the last 4 months. And this was despite the fact that both projects
already used TSC which would offer this functionality.

Currently, the upload functionality in TSC is hard to discover as it is not
exposed like all other REST functions. Instead of `server.fileuploads`, one
has to first create an instance of the (undocumented) `Fileuploads` class.

The upload functionality was probably because it should be usually unnecessary:
The uploaded files are usually part of publishing a workbook/datasource/...
and the corresponding `datasources.publish` (and similar) already take care of
the upload internally.
However, TSC isn't always up-to-date with new REST APIs, and by exposing file
uploads directly we can make sure to offer the best possible experience to
users of TSC also in those transition periods.

This commit:
* turns the `Fileuploads` class into a normal endpoint class which is not tied
  to one upload (So far, `Fileuploads` was not stateless. Now it is)
* adds the endpoint to `server`, such that file uploads are available as
  `server.fileuploads`
* adjusts all other users to use `server.fileuploads` instead of constructing
  an ad hoc instance of the `Fileuploads` class

Documentation will be added in a separate commit.

[1] https://github.com/jharris126/tableau-data-update-api-samples/blob/41f51ae4d220de55caf63e91fe9eff5694b9456a/basic/basic_incremental_load.py#L23
[2] https://github.com/tableau/hyper-api-samples/blob/382e66481ec8339407cf9cfa5d41fcdcf3f6a0fb/Community-Supported/clouddb-extractor/tableau_restapi_helpers.py#L165