Stop supporting Python 3.5
Python 3.5 is already end-of-life and no longer receives security patches for over a year now.
Also, I recently added an f-string because all Python versions which are still in support, also
support f-strings. I was unpleasantly surprised when I had to realize that we still claim to
support Python 3.5 which doesn't have f-strings...