feat: allow setting page_size in .all and .filter
1399 introduced a `with_page_size` method that allowed a user
to specify the page_size of requests in the chains. It felt
awkward in practice, so this moves it to be a keyword only
argument of the `.all` and `.filter` methods for querysets.

As 1399 has not yet been merged into master, this should be a non
breaking change for consumers of TSC.