fix: Pager typing
Pager Protocols were missing the generic flags. Added those in
so the Pager correctly passes through the typing information.
Also removes the kwargs from the function signature of the Endpoint.get
protocol to make the Workbook endpoint match.

Adding a return annotation of `None` to the tests is very important
because it is what enables static type checkers, like mypy, to inspect
those functions. With these annotations now in place, users of TSC
should more transparently be able to carry through typing information
when using the Pager.