fix: queryset support for flowruns (#1460)
* fix: queryset support for flowruns

FlowRun's get endpoint does not return a PaginationItem. This
provides a tweak to QuerySet to provide a workaround so all items
matching whatever filters are supplied.

It also corrects the return types of flowruns.get and fixes the XML
test asset to reflect what is really returned by the server.

* fix: set unknown size to sys.maxsize

Users may length check a QuerySet as part of a normal workflow. A len of
0 would be misleading, indicating to the user that there are no matches
for the endpoint and/or filters they supplied. __len__ must return a
non-negative int. Sentinel values such as -1 or None do not work. This
only leaves maxsize as the possible flag.

* fix: docstring on QuerySet

* refactor(test): extract error factory to _utils

* chore(typing): flowruns.cancel can also accept a FlowRunItem

* style: black

---------

Co-authored-by: Jordan Woods <13803242+jorwoods@users.noreply.github.com>