fix: handle 0 item response in querysets (#1501)
* fix: handle 0 item response in querysets

A flaw in the __iter__ logic introduced to handle scenarios
where a pagination element is not included in the response xml
resulted in an infinite loop. This PR introduces a few changes
to protect against this:

1. After running QuerySet._fetch_all(), if the result_cache is
empty, return instead of performing other comparisons.
2. Ensure that any non-None total_available is returned from
the PaginationItem's object.
3. In _fetch_all, check if there is a PaginationItem that has been
populated so as to not call the server side endpoint muliple times
before returning.

* fix: null out PaginationItem._page_number

Tests were failing because the fetch_all method added a second
check before fetching the next page. This fix will allow the
next page to be retrieved when used normally

---------

Co-authored-by: Jordan Woods <13803242+jorwoods@users.noreply.github.com>