feat: retrieve tableau server product name
Closes #1592

Calls a VizPortal API to retrieve the detailed product name, and
attaches it to the TSC.Server object which will make determining
what requests to build for things like subscriptions easier. Adds
this functionality into the pre-existing `use_server_version`
function because:

1. Users of the library will already know to call that method.
2. Users already aware that method makes API calls.

If the request errors or if the payload doesn't match the
expected format, defaults to assuming TSC is talking with an
on-prem Server instance.

Co-authored-by: emeric-dsj