feat: make refresh consistent between endpoints
Workbooks.refresh correctly accepted both an id string and a
WorkbookItem to handle triggering refreshes. Datasources.refresh
technically accepted both, but the type annotations did not reflect
that. Flows.refresh only accepted a FlowItem, but will now accept
a flow_id. This eliminates the need to create otherwise hollow
FlowItems to pass to the refresh method when you otherwise have
the id.