fix: datasource owner/project missing parsing (#1700)
Eric Summers pointed out that when the User Visibility setting
is set to "Limited," TSC fails to parse because it can't retrieve
any owner information. This bug is due to an `UnboundLocalError`
where the `owner` and `project` variables were not assigned in cases
where the owner and project elements were not included in the XML
response.

This PR also includes a test for the parsing where the owner and
project elements are missing and properly set to `None` on the
DatasourceItem.

Co-authored-by: Jordan Woods <13803242+jorwoods@users.noreply.github.com>