feat: handle safe list noop and delete appropriately
Conversation with jacksonlauren indicated that if the
ExtensionSiteSettings safe list is None, no safeList element should be
included in the request. If the ExtensionSiteSettings.safe_list is a
non-None Iterable that yields 0 items, an empty safeList element will be
included in the request, having the effect of deleting the safe list on
the site.